confirm al borrar

+++ src/views/ProfileView.vue
@@ -25,11 +25,11 @@ export default {
                 this.usuario = responseComapny.data
             }
         } catch (error) {
-            this.addMsgArray('danger','No se ha podido recuperar los datos intentelo mas tarde')
+            this.addMsgArray('danger', 'No se ha podido recuperar los datos intentelo mas tarde')
         }
     },
     methods: {
-        ...mapActions(useStore, ['cleanUser','addMsgArray']),
+        ...mapActions(useStore, ['cleanUser', 'addMsgArray']),
         edit() {
             if (this.user.rol === 'COMP') {
                 this.$router.push('/company-mod/' + Number(this.usuario.id))
@@ -40,16 +40,18 @@ export default {
         async eliminar() {
             const apiService = new APIService(this.user.token)
             try {
-                if (this.user.rol === 'COMP') {
-                    await apiService.deleteCompany(this.usuario.CIF)
-                } else if (this.user.rol === 'STU') {
-                    await apiService.deleteStudent(this.usuario.id)
+                if (confirm("¿Seguro que quieres eliminar tu perfil?")) {
+                    if (this.user.rol === 'COMP') {
+                        await apiService.deleteCompany(this.usuario.CIF)
+                    } else if (this.user.rol === 'STU') {
+                        await apiService.deleteStudent(this.usuario.id)
+                    }
                 }
                 this.$router.push('/')
                 localStorage.clear()
                 this.cleanUser()
             } catch (error) {
-                this.addMsgArray('danger','Problemas al intentar eliminar el perfil, vuelva a intentarlo')
+                this.addMsgArray('danger', 'Problemas al intentar eliminar el perfil, vuelva a intentarlo')
             }
 
         }
@@ -100,7 +102,8 @@ export default {
 .container {
     max-width: 70%;
 }
-.btn{
+
+.btn {
     margin-right: 15px;
 }
 </style>
